# imports/nomad-life/imports/keyswitch/spec/openapi.yaml
openapi: "3.0.2"
info:
  title: Keyswitch API
  version: "1.0.0"
  description: |
    Keyswitch is a compliant peer-to-peer vehicle rental platform focused on airport compliance, mid/long-term rentals, specialty inventory, and transparent risk/insurance.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development server
...
paths:
...
  /bookings:
    post:
      tags:
        - Bookings
      summary: Create a new booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                listingId:
                  type: string
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Booking created successfully
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT